Feature: Add route to manually bind a status list entry (#1687)
* feat: Add API route to expose assign_status_entries

Due to the nature of the status-list plugin, unless there is python code
elsewhere to assign status entries to credentials, there currently
exists no way to assign a status list entry to a credential. Because of
this fact, I have added an API that allows a controller to manually
assign credential status entries to a credential.

Signed-off-by: Colton Wolkins (Laptop) <colton@indicio.tech>

* fix: Fix route format for consistency

Signed-off-by: Colton Wolkins (Laptop) <colton@indicio.tech>

* fix: Add missing contect to function call

Signed-off-by: Colton Wolkins (Laptop) <colton@indicio.tech>

* fix: API Route path

Signed-off-by: Colton Wolkins (Laptop) <colton@indicio.tech>

* fix: schema validation

Signed-off-by: Colton Wolkins (Laptop) <colton@indicio.tech>

---------

Signed-off-by: Colton Wolkins (Laptop) <colton@indicio.tech>